+++ index.html
@@ -226,6 +226,15 @@
       </div>
     </section>
   </template>
+
+  <!-- Шаблон комментария -->
+  <template id="social__comment">
+    <li class="social__comment">
+      <img class="social__picture" src="" alt="Аватар комментатора фотографии" width="35" height="35">
+      <p class="social__text"></p>
+    </li>
+  </template>
+
   <script type="module" src="./js/main.js"></script>
   <script src="./js/functions.js"></script>
 </body>
