+++ js/fullscreen-mode.js
@@ -1,16 +1,25 @@
-import {isEscapeKey} from './util.js';
+import {onPopupEscKeydown} from './util.js';
+
+const commentsContainer = document.querySelector('.social__comments');
+const commentSample = document.querySelector('#social__comment')
+  .content
+  .querySelector('.social__comment');
+
+const renderComments = (comments) => {
+  commentsContainer.innerHTML = '';
+  comments.forEach(({avatar, name, message}) => {
+    const commentElement = commentSample.cloneNode(true);
+    commentElement.querySelector('.social__picture').src = avatar;
+    commentElement.querySelector('.social__picture').alt = name;
+    commentElement.querySelector('.social__text').textContent = message;
+    commentsContainer.append(commentElement);
+  });
+};
 
 const bigPicture = document.querySelector('.big-picture');
 const commentCount = bigPicture.querySelector('.social__comment-count');
 const commentsLoader = bigPicture.querySelector('.comments-loader');
 const body = document.querySelector('body');
-
-const onPopupEscKeydown = (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    closeBigPicture();
-  }
-};
 
 const openBigPicture = ({url, description, likes, comments}) => {
   bigPicture.classList.remove('hidden');
@@ -18,16 +27,10 @@ const openBigPicture = ({url, description, likes, comments}) => {
   bigPicture.querySelector('.likes-count').textContent = likes;
   bigPicture.querySelector('.comments-count').textContent = comments.length;
   bigPicture.querySelector('.social__caption').textContent = description;
-  const commentsList = bigPicture.querySelector('.social__comments');
-  commentsList.innerHTML = '';
-
-  comments.forEach((comment) => {
-    commentsList.append(comment);
-  });
   body.classList.add('modal-open');
+  renderComments(comments);
   commentCount.classList.add('hidden');
   commentsLoader.classList.add('hidden');
-
 
   document.addEventListener('keydown', onPopupEscKeydown);
 };
@@ -42,4 +45,3 @@ const closeBigPicture = () => {
 };
 
 export {openBigPicture, closeBigPicture};
-
