+++ js/main.js
@@ -1,4 +1,1 @@
-import {renderPhotos} from './similar-photos.js';
-import './photo-viewer.js';
-
-renderPhotos();
+import './gallery.js';
