+++ js/util.js
@@ -1,3 +1,5 @@
+import {closeBigPicture} from './fullscreen-mode.js';
+
 const getRandomInteger = (min, max) => {
   const lower = Math.ceil(Math.min(min, max));
   const upper = Math.floor(Math.max(min, max));
@@ -25,8 +27,13 @@ const createUniqueId = (min, max) => {
 
 const isEscapeKey = (evt) => evt.key === 'Escape';
 
-const getSelectedPhotoIndex = (array, evt) => {
-  return array.findIndex((obj) => obj.url === evt.target.getAttribute('src'));
+const onPopupEscKeydown = (evt) => {
+  if (isEscapeKey(evt)) {
+    evt.preventDefault();
+    closeBigPicture();
+  }
 };
 
-export {getRandomInteger, createUniqueId, getRandomArrayElement, isEscapeKey, getSelectedPhotoIndex};
+const getSelectedPhotoIndex = (array, evt) => array.findIndex((obj) => obj.url === evt.target.getAttribute('src'));
+
+export {getRandomInteger, createUniqueId, getRandomArrayElement, isEscapeKey, onPopupEscKeydown, getSelectedPhotoIndex};
